drivers: jtag_uart: use zephyr/posix prefix for posix headers
Approximately 2 and 1/2 years ago, Zephyr switched to requiring
a prefix of <zephyr/posix/..> for standard posix headers.

It would seem as though the altera hal is unmaintained and
nobody bothered to update it.

Note: It is advisable to not use POSIX within drivers since it
creates a dependency cycle, so this is still somewhat broken.

Signed-off-by: Chris Friedt <cfriedt@tenstorrent.com>